requirements.txt
infi.systray==0.1.12
plyer==2.1.0
psutil==5.9.8
pypiwin32==223
pywin32==306
setuptools==69.5.1
